requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
crispy-bootstrap5==2024.2
diff-match-patch==20230430
dj-database-url==2.2.0
Django==4.2.16
django-bootstrap-v5==1.0.11
django-crispy-forms==2.3
django-filter==24.3
django-heroku==0.3.1
django-import-export==4.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.8
Markdown==3.6
packaging==24.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.7.0
